separate amici objective and utils
